Support workflow_run triggered by push event workflows
Updates to enhance security around github token handling means that PRs
merged to a target branch by github native dependabot no longer receive
a write access token by default. Instead it is required to have a push
workflow trigger a workflow_run subsequently that is provided with write
access.

To support this add a route and handler function for workflow_run events
that reuses much of the functionality from the push handler by moving
the common capabilities to a shared function.

Includes a check to skip such a workflow if it is triggered via
pull_request events instead until such time as someone can add the
required support using the existing pull_request handler for the common
pieces.

Adds tests following the same pattern for the handlePull function.